LF-4378 Rename MultiStepForm
