+++ packages/webapp/src/stories/ContextForm/ContextForm.stories.tsx
@@ -15,17 +15,17 @@
 
 import { Meta, StoryObj } from '@storybook/react';
 import { componentDecorators } from '../Pages/config/Decorators';
-import { MultiStepForm, VARIANT } from '../../components/Form/MultiStepForm';
+import { ContextForm, VARIANT } from '../../components/Form/ContextForm';
 
 // https://storybook.js.org/docs/writing-stories/typescript
-const meta: Meta<typeof MultiStepForm> = {
-  title: 'Components/MultiStepForm',
-  component: MultiStepForm,
+const meta: Meta<typeof ContextForm> = {
+  title: 'Components/ContextForm',
+  component: ContextForm,
   decorators: componentDecorators,
 };
 export default meta;
 
-type Story = StoryObj<typeof MultiStepForm>;
+type Story = StoryObj<typeof ContextForm>;
 
 export const PageTitle: Story = {
   args: {
